+++ build.gradle
@@ -33,6 +33,16 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+
+	ext {
+		JJWT_VERSION = "0.12.5"
+	}
+
+	implementation "io.jsonwebtoken:jjwt-api:${JJWT_VERSION}"
+	runtimeOnly "io.jsonwebtoken:jjwt-impl:${JJWT_VERSION}"
+	runtimeOnly "io.jsonwebtoken:jjwt-jackson:${JJWT_VERSION}"
 }
 
 tasks.named('test') {
