+++ src/main/java/com/pet/pet_funeral/domain/user/dto/KakaoLoginRequest.java
@@ -1,0 +1,4 @@
+package com.pet.pet_funeral.domain.user.dto;
+
+public class KakaoLoginRequest {
+}
