+++ src/main/java/com/pet/pet_funeral/domain/user/dto/KakaoTokenResponse.java
@@ -1,8 +1,7 @@
-package com.pet.pet_funeral.kakao.dto;
+package com.pet.pet_funeral.domain.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
