+++ src/main/java/com/pet/pet_funeral/domain/user/dto/KakaoUserResponse.java
@@ -1,11 +1,9 @@
-package com.pet.pet_funeral.kakao.dto;
+package com.pet.pet_funeral.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.UUID;
 
 @Getter
 @AllArgsConstructor
