+++ src/main/java/com/pet/pet_funeral/domain/user/model/Role.java
@@ -1,6 +1,6 @@
 package com.pet.pet_funeral.domain.user.model;
 
 public enum Role {
-    ROLE_ADMIN,
-    ROLE_USER
+    ADMIN,
+    USER
 }
