+++ src/main/java/com/pet/pet_funeral/domain/user/repository/RefreshTokenRepository.java
@@ -7,8 +7,9 @@ import org.springframework.stereotype.Repository;
 
 
 import java.util.Optional;
+import java.util.UUID;
 
 @Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, UUID> {
     Optional<RefreshToken> findByToken(String token);
 }
