--- src/main/java/com/pet/pet_funeral/kakao/dto/KakaoLoginRequest.java
@@ -1,4 +1,0 @@
-package com.pet.pet_funeral.kakao.dto;
-
-public class KakaoLoginRequest {
-}
